♻️ refactor(recipes): fix use-case class casing Usecase -> UseCase
Rename use-case classes (and their files + references) from the
inconsistent "...Usecase" form to PascalCase "...UseCase".
